+++ src/main/java/com/example/secondlife/domain/comment/service/CommentSearchService.java
@@ -1,41 +1,27 @@
-package com.example.secondlife.common.service;
+package com.example.secondlife.domain.comment.service;
 
 import com.example.secondlife.domain.comment.dto.CommentResponse;
 import com.example.secondlife.domain.comment.entity.Comment;
 import com.example.secondlife.domain.comment.repository.CommentRepository;
-import com.example.secondlife.domain.post.entity.Post;
-import com.example.secondlife.domain.post.repository.PostRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Slf4j
-public class PostCommentService {
+@Transactional(readOnly = true)
+public class CommentSearchService {
 
-    private final PostRepository postRepository;
     private final CommentRepository commentRepository;
 
-    @Transactional(readOnly = true)
     public List<CommentResponse> getComments(Long postId) {
         List<Comment> comments = commentRepository.findByPostId(postId);
 
         return Comment.toCommentResponseList(comments);
     }
 
-    @Transactional(readOnly = true)
-    public Post findPostById(Long postId) {
-        log.info("findById");
-
-        return postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 존재하지 않습니다. postId = " + postId));
-    }
-
-    @Transactional(readOnly = true)
-    public Comment findCommentById(Long commentId) {
+    public Comment findById(Long commentId) {
         return commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 댓글이 존재하지 않습니다. commentId = " + commentId));
     }
