+++ src/main/java/com/example/secondlife/domain/comment/service/CommentService.java
@@ -1,11 +1,11 @@
 package com.example.secondlife.domain.comment.service;
 
-import com.example.secondlife.common.service.PostCommentService;
 import com.example.secondlife.domain.comment.dto.CommentRequest;
 import com.example.secondlife.domain.comment.dto.CommentResponse;
 import com.example.secondlife.domain.comment.entity.Comment;
 import com.example.secondlife.domain.comment.repository.CommentRepository;
 import com.example.secondlife.domain.post.entity.Post;
+import com.example.secondlife.domain.post.service.PostSearchService;
 import com.example.secondlife.domain.user.entity.User;
 import com.example.secondlife.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -19,11 +19,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
     private final CommentRepository commentRepository;
-    private final PostCommentService postCommentService;
+    private final PostSearchService postSearchService;
+    private final CommentSearchService commentSearchService;
     private final UserService userService;
 
     public CommentResponse save(Long postId, Long userId, CommentRequest request) {
-        Post findPost = postCommentService.findPostById(postId);
+        Post findPost = postSearchService.findById(postId);
         User findUser = userService.findById(userId);
 
         Comment comment = request.toEntity(findPost, findUser);
@@ -34,7 +35,7 @@ public class CommentService {
     }
 
     public CommentResponse updateComment(Long commentId, Long userId, CommentRequest request) {
-        Comment findComment = postCommentService.findCommentById(commentId);
+        Comment findComment = commentSearchService.findById(commentId);
 
         validUser(userId, findComment);
 
@@ -44,7 +45,7 @@ public class CommentService {
     }
 
     public void deleteComment(Long commentId, Long userId) {
-        Comment findComment = postCommentService.findCommentById(commentId);
+        Comment findComment = commentSearchService.findById(commentId);
 
         validUser(userId, findComment);
