+++ src/main/java/com/example/secondlife/domain/post/entity/Post.java
@@ -1,7 +1,9 @@
 package com.example.secondlife.domain.post.entity;
 
 import com.example.secondlife.common.base.BaseEntity;
+import com.example.secondlife.domain.comment.dto.CommentResponse;
 import com.example.secondlife.domain.post.dto.PostRequest;
+import com.example.secondlife.domain.post.dto.PostResponse;
 import com.example.secondlife.domain.post.enumType.Forum;
 import com.example.secondlife.domain.user.entity.User;
 import jakarta.persistence.Column;
@@ -14,6 +16,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -69,6 +72,31 @@ public class Post extends BaseEntity {
         this.hits = hits;
         this.isPublic = isPublic;
         this.forum = forum;
+    }
+
+
+    public PostResponse postWithCommentToPostResponse(List<CommentResponse> comments) {
+
+        PostResponse postResponse = toPostResponse();
+        postResponse.setCommentResponse(comments);
+        return postResponse;
+
+    }
+
+    public PostResponse toPostResponse() {
+
+        return PostResponse.builder()
+                .userId(user.getId())
+                .postId(id)
+                .title(title)
+                .contents(contents)
+                .hits(hits)
+                .createdDate(getCreatedDate())
+                .lastModifiedDate(getLastModifiedDate())
+                .createdBy(getCreatedBy())
+                .lastModifiedBy(getLastModifiedBy())
+                .build();
+
     }
 
 }
