+++ src/main/java/com/example/secondlife/domain/post/service/PostService.java
@@ -1,18 +1,13 @@
 package com.example.secondlife.domain.post.service;
 
-import com.example.secondlife.common.service.PostCommentService;
-import com.example.secondlife.domain.comment.dto.CommentResponse;
 import com.example.secondlife.domain.post.dto.PostRequest;
 import com.example.secondlife.domain.post.dto.PostResponse;
 import com.example.secondlife.domain.post.entity.Post;
 import com.example.secondlife.domain.post.repository.PostRepository;
 import com.example.secondlife.domain.user.entity.User;
 import com.example.secondlife.domain.user.service.UserService;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,53 +20,32 @@ public class PostService {
 
     private final PostRepository postRepository;
     private final UserService userService;
-    private final PostCommentService postCommentService;
+    private final PostSearchService postSearchService;
 
     public PostResponse save(Long userId, PostRequest request) {
         log.info("save");
 
         Post savedPost = postRepository.save(postRequestToPost(userId, request));
 
-        return postToPostResponse(savedPost);
+        return savedPost.toPostResponse();
     }
-
-    @Transactional(readOnly = true)
-    public Page<PostResponse> getPosts(Pageable pageable) {
-        log.info("getPosts");
-
-        Page<Post> posts = postRepository.findAll(pageable);
-
-        return posts.map(this::postToPostResponse);
-    }
-
-    @Transactional(readOnly = true)
-    public PostResponse readWithComments(Long postId) {
-        log.info("readWithComments");
-
-        Post post = postCommentService.findPostById(postId);
-
-        List<CommentResponse> comments = postCommentService.getComments(postId);
-
-        return postWithCommentToPostResponse(post, comments);
-    }
-
 
     public PostResponse updatePost(Long userId, Long postId, PostRequest request) {
         log.info("updatePost");
 
-        Post findPost = postCommentService.findPostById(postId);
+        Post findPost = postSearchService.findById(postId);
 
         validUser(userId, findPost);
 
         findPost.update(request);
 
-        return postToPostResponse(findPost);
+        return findPost.toPostResponse();
     }
 
     public void deletePost(Long userId, Long postId) {
         log.info("deletePost");
 
-        Post findPost = postCommentService.findPostById(postId);
+        Post findPost = postSearchService.findById(postId);
 
         validUser(userId, findPost);
 
@@ -100,27 +74,4 @@ public class PostService {
 
     }
 
-    public PostResponse postWithCommentToPostResponse(Post post, List<CommentResponse> comments) {
-
-        PostResponse postResponse = postToPostResponse(post);
-        postResponse.setCommentResponse(comments);
-        return postResponse;
-
-    }
-
-    public PostResponse postToPostResponse(Post post) {
-
-        return PostResponse.builder()
-                .userId(post.getUser().getId())
-                .postId(post.getId())
-                .title(post.getTitle())
-                .contents(post.getContents())
-                .hits(post.getHits())
-                .createdDate(post.getCreatedDate())
-                .lastModifiedDate(post.getLastModifiedDate())
-                .createdBy(post.getCreatedBy())
-                .lastModifiedBy(post.getLastModifiedBy())
-                .build();
-
-    }
 }
